+++ modules/term/src/test/groovy/com/intellisrc/term/TableMakerTest.groovy
@@ -1,5 +1,6 @@
 package com.intellisrc.term
 
+import com.intellisrc.core.Log
 import com.intellisrc.term.styles.*
 import spock.lang.PendingFeature
 import spock.lang.Specification
@@ -370,5 +371,40 @@ class TableMakerTest extends Specification {
             tp.print()
         then:
             assert tp.toString().trim() == answer17
+    }
+
+    def "Map should print nicely"() {
+        setup:
+            Map map = [
+                name        : "Samantha Wigs",
+                age         : 25,
+                country     : "New Zealand",
+                occupation  : "Teacher",
+                email       : "sam25@teachers.nz",
+                phone       : "098-8712-9378"
+            ]
+            TableMaker tm = new TableMaker(map, horizontal)
+        when:
+            tm.print()
+        then:
+            assert tm.toString().trim() == answer
+        where:
+            horizontal | answer
+            false      | answer18
+            true       | answer19
+    }
+
+    def "Must handle empty data without errors"() {
+        setup:
+            TableMaker tm = new TableMaker([])
+        expect:
+            try {
+                tm.print()
+                assert true
+            } catch(Exception e) {
+                Log.e("Unable to print", e)
+                assert false : "Error reported"
+            }
+
     }
 }